moved optimize_trajectories into Morris class and cleaned up tests. All of the Morris helper functions are now in morris_util.py, separate from the main class. Still need to figure out how to test the group/optimal sampling functions which are now inside the main class.
